ANDROID: Do not submit. Change to test workflow.
* DO NOT SUBMIT
* DO_NOT_SUBMIT

This change is intended just for testing purposes.
To test the ABI presubmit workflow.

Bug: 243136716

+++ arch/arm64/kernel/pci.c
@@ -56,7 +56,7 @@ int raw_pci_write(unsigned int domain, unsigned int bus,
 
 #ifdef CONFIG_NUMA
 
-int pcibus_to_node(struct pci_bus *bus)
+long pcibus_to_node(struct pci_bus *bus)
 {
 	return dev_to_node(&bus->dev);
 }
